Clean up airs duplicate table IDs
Both sms.json and sms_case.json have a table_id of static-sms-indicators-001
Their sets of domains don't overlap at all, which is handy.

sms.json only references airsghana and airszimbabwe.  airsghana has no reports
which use static-sms-indicators-001, however, so that's useless.  airszimbabwe
does have many reports which reference this DS, but they are all reports with a
bunch of other domains for which the DS is sms_cases.  I find it extremely
unlikely that these reports are expected to work on both a case data source for
some domains and a form data source for another, and in fact these reports fail
to load on production (or do load, but with no data).

Ergo, I'm removing sms.json entirely, and adding airszimbabwe to sms_cases.json,
which removes the ambiguity around duplicate data sources, and should fix these
reports for airszimbabwe.