Make context manager to clean tempfiles
This is intended to establish a pattern for accessing tempfiles, where there's a
with statement for file reading (like normal), and a with statement for the
existence of the file.

tempfile.TemporaryFile does automatically clean itself up on close, but then if
you're planning to both read and write to the file, you need to `seek` around.
This wrapper lets us open for reads and open for writes like we normally do,
but still have an automatic cleanup step.

One potential issue with this is that you can apparently remove a file without
first closing it.  That could cause weird bugs if something went wrong.  This
util could control file access as well, but that might be overly defensive.  I'm
sticking this approach for now to keep things simple and preserve the standard
python API as much as possible.