Add recovery measures url to app profile
Note that it looks like a lot of this code expects the profile ID to be missing
sometimes.  If that's possible, it'd probably cause a 404 when the phone tries
to hit the recovery endpoint.

There are several existing URLs in here which should suffer from the same
issue (eg heartbeat, profile, media profile), so I'm guessing that's unlikely to happen.