Invalidate data definition class
This was causing an intermittent test failure.

A regrettable side effect of doing this was that I had to drop the import
statements from custom_data_fields.__init__.  This appears to be because of this
change from Django 1.9:

> All models need to be defined inside an installed application or declare an
explicit app_label. Furthermore, it isn’t possible to import them before their
application is loaded. In particular, it isn’t possible to import models inside
the root package of an application.