Make columns available for in-form filtering
There are several dozen mobile reports configured in the LS app (and I believe
more in the AWW app) which apply filters to reference a subset of report data.
The app then dynamically selects which report config to use based on user input.
This dynamism isn't really possible with mobile UCR v2 (at least, not simply).

For example, there are 6 reports based on the report
"MPR - 6b - Child Health cases (Static) v2":

    child cases age 36-48 months, current month
    child cases age 48-60 months, current month
    child cases age 60-72 months, current month
    child cases age 36-48 months, previous month
    child cases age 48-60 months, previous month
    child cases age 60-72 months, previous month

The approach we're looking at is to deduplicate these reports by making
the requisite columns available in the restore, so they may be filtered
within the app.