Always assign new form id in same shard as original
Prior to this fix an unexpected error during form submission (see couchforms.tests.test_edits:EditFormTestSQL.test_broken_save) could cause a new form to be saved with a bogus deprecated_form_id. In this case, because the form was not saved at time of handling the unexpected error, the form was assigned a new/random UUID, which could land in any shard (and may differ from its deprecated_form_id shard). Since deprecated_form_id is not null the form would later try to reparent deprecated attachments before writing new attachments, which resulted in a shard match assertion failure.

This fix seems to be easiest work around, and I don't think it will do any harm. However, it is less than ideal that a form is being saved with bogus deprecated_form_id. It would be nice to avoid that situation, but that looks more complicated.