Bump form timeout to 15 minutes
Hoping this will alleviate this:
https://manage.dimagi.com/default.asp?269112

Some supply updates are being applied twice.  The suspicion is that they
take so long to process that they are hitting a timeout and being
submitted/processed a second time.  More context in that ticket.

Here's a branch which attempts to reproduce the issue:
https://github.com/dimagi/commcare-hq/compare/es/duplicate-ledgers
(commit 9b4eadf if that's gone)