Removed angular tests.
Most of the tests just check that the js variables are set to the values
passed back by the initial ajax call. It would be good to have tests of
the drilldown interactivity, but these aren't that, and creating those
right now is prohibitively time-consuming because the drilldown is
based on a django form.