Wrap instantiation of ParentCasePropertyBuilder
There are two distinct usages of this, and a lot of duplicated setup.
This approach should constrain each a bit and simplify the calling code.
Note that there _is_ a change here - not all of these instances provided
the per_type_defaults, but they really don't affect much and aren't
expensive to compute.