Fixed conflation of user_types and userTypes.
Confusingly, this file uses "user types" in two ways. One is that the export
download form has a legacy user_types field, which was removed in
https://github.com/dimagi/commcare-hq/pull/22141/commits/41cb9a6b2b62547a224fe9d611bb749f1c001c09
The other is a list of human-readable names for the major categories of mobile workers,
which is used to send a human-readable string to analytics instead of a string like "t__0,t__4".
This commit removes the first usage and makes some stylistic adjustments to the second.