Provide document_id as key for kafka partitioning
This will ensure that forms and cases reliably go to the same partition
when saved.

It provides a potential speedup in the event that the same document is
saved multiple times rapidly and we can de-duplicate its processing
(such as a pillow that processes changes in bulk). It also reduces the
likelihood that multiple processes will process the same document at the
same time.

The document will not be guaranteed to be in the same partition when we
add partitions to kafka.