Removed icds_rebuild_related_docs from data sources
This was originally implemented to handle large discrepancies between
the speed of processing ccs_record and child_health cases on the
dashboard. During that time we had a very large (> 10 day) lag in
processing the data sources.

Since then we have totally removed child_health's "tableau" data source
and greatly improved performance, so this is no longer doing anything
useful.