Redirect form's source to release's page
Before this commit, a user was able to modify a build's form's XML in
place. It's unclear if that has affected any applications before as I
have not investigated how much CCZs are cached or how often a build's
URLs are hit.

`view_generic` already does this for a majority of app builder URLs.