Add back in checks for the domain object
While I'm still pretty sure this doesn't happen in production, it _does_
apparently happen a lot in tests (users assigned to a domain that
doesn't exist in the db).  Turns out this has been failing all over the
place, but because it's in a signal, those failures were silenced.  Yay
signals!