Upgraded InvoiceInterfaceBase
This one is a little weird: I'm switching over to use select2 v4
even though this is a reporting interface, so it appears on a reporting
page, which is still on legacy select2. This works okay because the
none of the select2s in accounting/js/widgets are used; this page only
depends on that file to initialize the adjust balance form. This would
not work if reports were on requirejs - so upgrading select2 in reports
is a prerequisite for migrating them to requirejs.