Change != to IS DISTINCT FROM in agg_awc aggregation
After reviewing person_cases_v3 data source, I realized that the only
values will be NULL or 0/1 for registered_status or migration_status.
This is because the app only stores whether a person is not registered
or migrated and a NULL value if its unknown. I have changed our fixture
data source to respect that convention and used IS DISTINCT FROM and !=
to better adhere to the convention that person_cases_v2 data source had
set in place