Simplify to/from calculation
This wasn't returning proper values in some instances, and to be honest
I didn't really understand what exactly it was trying to do.  This new
version pulls the "from" from the element itself, since we already have
it, and it calculates "to" based on the index of the element with that
UID in the target sortable.  Previously it looked for the element in the
target sortable with a different module UID.