Make celery check accept any process responds to any out of 20 pings
I am not sure this will help, but instead of doing five pings with a timeout of 10s,
and requiring one of them to have a perfect record (all processes respond),
do up to 20 pings with a timeout of 1s and accept any process that
responds to _any_ of those pings. Also ding a process that is supposed
to be stopped only if it responds to _all_ pings.