Migrated ews-ghana custom user settings page.
/a/DOMAIN/settings/users/web/account/USER_ID/

This page doesn't work on prod: the select2 throws an error.
Analytics say it hasn't been used since Oct 2017. I'm following up on
the possibility of removing it. In the meantime, this switches it from
a v3 select2 that throws an error to a v4 select2 that throws an error.