Have initiate_queue_scheduled_reports call queue_scheduled_reports directly
Currently initiate_queue_scheduled_reports calls queue_scheduled_reports.delay,
which puts it back on the periodic_queue. This means that we need to wait
two round trips on the periodic_queue before we even start running any code
related to scheduled reports. This cuts it down to a single round trip on the periodic_queue.