Start per-queue "uptime" as commcare.celery.heartbeat.blockage_ok
This is a binary version of the blockage_duration metric
and reports only whether or not the blockage_duration
is under the particular threshold for each queue.

This allows us to construct individual uptime graphs per queue
and is a stronger signal for how well we're doing per queue
than blockage_duration, which has different "normal" values
for every queue.