Add ucr tables to domain delete operations
I think this is a safe way to do this. First, pull all tables in the ucr db that
have the domain name in them, just as a way to filter down to a very small set
to pull into memory. Then, for each one, pull out the table prefix and table id,
reconstruct the expected table name from those, and only keep the ones
where the original table name matches the expected table name.

I think I can assert that this will _never_ match a table in the wrong domain.
For a very long domain name, however, this may _miss_ some tables in the domain,
specficially if the table name was trucated so that the full table id is not present
in the table name. Then I think short of some sort of brute force search,
there is no simple way to show that a table belongs to a domain.