Fix case attachments diff test
This test is relatively new, and the case attachment keys were chosen arbitrarily based on the previously-existing diff rules and comments, some of which were wrong. The "couch only" comment on "content_type" was wrong—it's actually always present for SQL and may be present in couch, although may be "server_mime" in couch.