Fix off-by-one error
This introduces a race condition.  The celery task returns a task status
with a result detailing the outcome of the upload, which is fetched and
saved immediately once the task is marked as complete, either by the
polling request on the case importer page or by a separate save_result
task kicked off at the end of the import task.  With this off-by-one
error, the task result briefly has `None` in place of the outcome JSON,
which is then saved to the upload.  Introducing a delay fixes that.

I'm not sure why fixing this progress info resolves that issue, but it
does.  Perhaps something along the way sees progress>total and declares
the task done premptively?