Fix DeID daily saved and dashboard exports
The DeID-only filtering now happens in the dbaccessor, so it can be dropped.
Also, these called `super` on the wrong class, which is a hard failure
in Python.  I was able to trigger that on production.  This bug has been
in place for 5 months, but only one other person had triggered it, so
perhaps we don't actually need DeID saved or dashboard exports