Fix resumable migration
The presence of this line is the result of a bad merge/rebase in https://github.com/dimagi/commcare-hq/pull/22272/files#diff-c730c38ab6f59f86107f0c58d1ce8e86R138
Note that the original PR removed that line https://github.com/dimagi/commcare-hq/pull/21007/files#diff-c730c38ab6f59f86107f0c58d1ce8e86L117

Part of the reason this slipped through was that there are no tests for the resume feature (it's hard to test because it's tightly coupled with the concurrency mechanism). I'm not sure if it was ever used, but if it was it could have resulted in some forms not being migrated. They should have been caught in the final diff stage of the migration.