Add and test efficient sliding window rate counter
It works by breaking the window into "grains" with fixed period
boundaries, and then treating the traffic from the earliest grain that's
sliding out of the window as if it were even over the period. This
trades off the accuracy of keeping the date of every event and the
efficiecy of counting within fixed time buckets, and with enough grains
it approaches the former.