Make conditionally_location_safe importer appear
https://dimagi-dev.atlassian.net/browse/PF-389
This makes the link to the case importer appear for location-restricted
users on domains with the feature flag LOCATION_SAFE_CASE_IMPORTS
enabled.

When uitabs checks if a URL is location safe, it does not pass `request`
to the conditional function (since request params will vary).  This only needed
access to domain, which is already available in `kwargs`, so a switch to
that makes it work.