Coerce falsy values to null
https://dimagi-dev.atlassian.net/browse/HI-716
Fixes an issue where changing module settings would trigger a reorder
This happened only with a particular combination of feature flags and
addons.  In particular, BASIC_CHILD_MODULE had to be disabled.  At any
rate, the issue was that `oldRoot` was `""` and `newRoot` was
`undefined`, since the the select parent widget was only available for
advanced modules.  Since `"" !=== undefined`, this incorrectly triggered
a reorder.