Rename --log-dir to --state-dir
It is not safe to store (SQLite db) state on an NFS volume when it will be accessed by multiple processes concurrently (feature pending). Therefore it is no longer a good idea to use the file system of settings.SHARED_DRIVE_CONF.tzmigration_planning_dir as a place to store migration state.