Track submissions that would be rate limited
Get the system in place for rate limiting submissions, but for now
instead of actually blocking usage, just report to datadog that a
submission would have been blocked.

The rate limiting is based on the domain-wide per-user rate, and is
given both a minimum usage to allow (e.g. for a domain with very few
users) and a per user rate for projects that have enough users.