Fix a bit of jankiness in the builds menu page
The cache was only cleared when the page was initially loaded, instead
of after changes were made.  This meant it wouldn't be reflected
elsewhere on the site unless you refreshed.  Also, since it didn't do a
redirect on successful POST, refreshing would attempt to resubmit the
same form, which would give a resource conflict.