Added filters to query string for UCR urls.
Not directly related to these tickets, but saved reports seem not to work for UCRs on prod.
Clicking a UCR link on the reports home takes you to the report, but with its default filters
instead of your saved filters.

I don't love just removing an if statement without knowing why it was added,
but git history was disrupted by a code move.