Fix sms calculated properties on domain
Was passing "I" and "O" to elasticsearch for the `direction` field,
but in elasticsearch the field has values "i" and "o", so the queries
were all returning 0. This changes to using the INCOMING and OUTGOING
constants (which are just "I" and "O") for passing values around and
the built in SMSES filters for incoming_messages and outgoing_messages
to fix the issue while also improving readability and reducing confusion.