Raise AttachmentNotFound on not saved
If an Attachment is saved and get_attachment() is called with
attachment_name set to a missing attachment, AttachmentNotFound will be
raised by get_attachment_meta(). But id the Attachment is not saved
get_attachment_meta() will return None, and attachment.open() will
cause an AttributeError. This change resolves that by raising the same
exception for missing attachments whether the object is saved or not.