Restrict updates to app submission metadata to the latest form
Beginning with https://github.com/dimagi/commcare-hq/pull/16219/ we
began to update the reporting data when the build or version was
changed, but this causes the last submission time to go backwards if we
reprocess forms at a later time.

In this case (https://dimagi-dev.atlassian.net/browse/ICDS-882) a user
upgraded her version of the app. Due to an infrastructure issue we
needed to reprocess a large amount of forms from a couple of weeks in
the past. When our pillow reprocessed it, it caused her last
submission to go from the past day to two weeks ago because the version
of the app was not the same.