Fix single case api, broken by ES refactor
Broken in https://github.com/dimagi/commcare-hq/pull/25743. That stopped
saving the _id field to the ES doc source, instead adding it back in
from metadata in a wrapper. This was necessary for compatability with ES
2 because ES 2 no longer allows haivng a _id field in doc source. The
issue was that in this one place we were accessing the doc directly
instead of through the wrapper, which caused the _id field to be
missing.