Remove ordering by doc_id
This was originally implemented to avoid conflicts where there were more
than BATCH_SIZE records with the same inserted_at, but that's highly
unlikely since we have microsecond accuracy on inserted_at.

We also have an equality statement that raises an error if the latest
processed date time == the time passed into the function.