Added update_form_unique_ids back to scrub_source
This gets form unique ids updated on app copy or import, so that
they will be across domain, not just app. They need to be unique
across domain for multimaster, so that when a linked app pulls
from different masters, the xform resource ids in suite.xml are
different for different masters, so that mobile updates properly.