handle routing based in 'instance' hint
This hint is used by Django. When you create an object that
has a related object e.g. a CaseTransaction with a CommCareCaseSQL,
Django will call the router to get the DB for the one object
based on the DB for the other.

This used to work before because the router would return
the proxy DB which would get overwritten in the save function