Added content for translator file
This file contains a row for each file the translator needs to generate:
one row per unique text string, if there is not already an audio file
for that text string in the app. The workbook has two sheets: one with just
text and file path, and a second with the same columns as the bulk app translation
download.