Attempt to fix too many open files error
This does not fix the error, but does better cleanup. The error is not a major concern since it is not likely to be seen in the scenarios where this is expected to be used. It may be necessary to revisit if parallel.Pool is used in a long-running process that creates many child processes.