Disable new setup of sms/call as a user's default two factor method
Users already using sms/call as their default two factor method
will continue to be able to use it. Any first time setup
or setup after resetting two factor auth will not have sms or call
 as an option.

This is both because call and SMS are widely not considered
sufficiently secure to act as two factor methods,
and because it is a fraud vulnerability.