Give RateLimiter.wait a windows_not_to_wait_on param
This will let us pause and retry requests that are throttled due to
their minute or second window for a few seconds in the hopes to get them
through in the same request, while rejecting requests that are throttled
due to their hour, day, or week window outright.

This will give a better experience for API users who infrequently submit
bursts of forms, while giving our system the benefit of spacing them
out and not having to handle the overhead of repeated retries for these
kinds of requests.