Attempt not to 429 burst submissions if a delay will do
When receiving a request that would be throttled by the second or
minute, wait and retry for up to 15 seconds. This is less load on the
server than rehandling client retries, and a better experience for the
user. This is especially true in the case of API usage, where the script
may not handle 429s well.