Unblock saved exports by detecting missing/expired task
For each saved export we keep track of the active task id
so that we do not have more than one task running for it at a time.
However, we recently found a few of them that were pointing to tasks
that do not appear to actually exist, and that put it in a stuck state
of always waiting for a task to finish that never will because it does not exist.
We now detect that and designate the task as "missing" rather than "not started".