New algorithm for SQL transaction sorting
SQL transactions are now sorted by the first appearance of the corresponding form id in Couch actions. Previously a SQL transaction sorting function similar to the the one used for Couch actions was used, but it could result in bad output if there were multiple couch actions per SQL transaction (due to multiple action types, which are combined in SQL).

Also in this change: changes to the rebuilt SQL case are not saved if there are still diffs after the rebuild.