Reset postgres and couchdb connections
When gipc forks processes it does not close non-gipc related sockets (https://github.com/jgehrcke/gipc/issues/22). In practice this meant that the parent and child processes shared connections. This was observed with couchdbkit, which sometimes returned the wrong results from a view call.

Postgres connections are supposedly vulnerable to the same issue (https://stackoverflow.com/q/40482555/10840), but that has not been observed.