Handle duplicate stock transactions
Some couch ledgers have duplicate stock transactions as a result of a bug that may still exist or it existed at one time and has subsequently been fixed. This commit addresses a specific known subset of that problem scenario where a ledger balance is wrong due to the duplicate stock transaction. There may be other more complicated manifestations of this bug that are not handled here. For example, when there are more than two transactions in the ledger history.