Fixed bug in check_for_missing_content
At the point where this is called, messages may contain a mix of
translated and untranslated content. This is fine in the
TranslatedConditionalAlertUploader implementation, which just checked
for blank values, but the UntranslatedConditionalAlertUploader was
looking specifically for the key '*' and would break if it received a
translated message, which woould only have language code keys.