Restore __hash__ on Subscription
I think even better would be to rename the method now called __eq__
to a normal method (e.g. "is_equivalent_to") that better captures what
it is doing. It doesn't check if it's the same db object like model
normally does, but rather check to see if it is an equivalent
subscription, meaning that two different rows in the db can have
row1 == row2 be True, which is surprising behavior.