Merge together FoodRow classes
I'm on the fence about this commit.  When I originally split up the
class into 3, I thought row from fixture vs row from ucr would be the
biggest division, and it is, but it's not actually that big a divide.

There's also recipes vs ingredients, standard recipes and ingredients
versus nonstandard ones, thingswith food composition table entries, and
so forth.

Anyways, given the mental overhead of maintaining a class hierarchy, I
thought this might be a bit easier to reason about.  I'm also open to
feedback and alternative suggestions.