When revoking a saved export task, also terminate
When we do this we expect the task is missing,
but revoke in in case we are wrong. Also in that case
we want to terminate it if it is indeed running.
If we make a mistake in determining it is missing,
we should at least commit to leaving things in a clean
state afterwards.