[Django 2] Surgically rearrange SQLInvitation db migrations
0011 leaves SQLInvitation in a state that is illegal in Django 2,
but that Django 1 was silently accepting.
Originally we tried squashing the last four migrations
(https://github.com/dimagi/commcare-hq/pull/27114\)
but there was then no reasonable way to sequence the
custom data migration part.

So instead we are surgically merging the 0011 back into
the original migration that created SQLInvitation
in a way that still leaves it workable for
the custom migration code which runs during 0010