Handle missing problem forms properly
There are three code paths on which forms are migrated, and two of them (on which missing forms are migrated) did not detect and transform problem forms into error forms. This resulted in some error forms being migrated as if they were normal forms, which means case and ledger updates were applied when they should not have been.