Fix the way we attempt to clear a stale channel claim
Previous commit referenced undefined name "session",
but that also prompted me to rethink this.
From a robustness perspective, we want to clear the _channel_
to be reclaimed whether or not the closed session for that stale claim
still is configured to point to that channel.
Introduces XFormsSessionSynchronization.clear_stale_channel_claim
for this purpose, with tests