enforce rule to not allow reusing usernames from deleted users
Since the introduction of the backend changes to warn users when reusing
a username from a deleted user, the UI has not permitted the creation
of those users (even though the backed would allow it).

This commit changes the backed to respond with an error message if
the username of a user who has been deleted is used to create a new user.