Make selected_gap_type filter rows not presence of gaps
Previously, selected_gap_type would filter rows to only those with gaps of that
type (as that's what that filter means elsewhere).  This report has a two rows
for every master row, one per gap type.  The filter here should control which of
those roles to return, and _not_ whether that gap_type is `AVAILABLE`