Added check for request.project back to base.html
I switched this earlier to check for request.secure_session, so that the popups
can show in domains that aren't secure (necessary if the user is a member of a
different, secure domain). But the domain check is still necessary, because the
login page used in the popup is the domain-specific login.