Fix project health report locally and in Django 2.2
There was an {% include %} tag in the template that has been failing silently in production
for probably years, resulting it just nothing being included.
Locally (DEBUG = True) this resulted instead in a 500, and in Django 2.1 or higher it
also results in a 500.

Thus this change essentially ports forward the existing behavior that the include tag get ignored
by deleting it altogether.