Switch flagging approach for HQ reports
The previous approach worked by updating a global when the report class
was first loaded, but that wasn't reliable.  Not all report classes are
guaranteed to have been loaded.  The major downside of this new approach
is that the safety flag is now inherited.  The same behavior exists for
Django CBVs though, so it seems a reasonable tradeoff.