Replaced CompositeView in BreadcrumbView
This one is convoluted. The desire for semantic markup - a single list
containing the static home link plus all subsequent links - is at odds
with Backbone wrapping each view in an element. Ideally there would be
one <ol> with a static home icon and then the CollectionView would
render the remaining list items.