Nit: Using `get()` isn't safer here
Using `get()` here doesn't avoid an error. If the key 'user_data' is not
found, this would just cause an AttributeError 3 or 7 lines later.

I chose `['user_data']` over `get('user_data', {})` because this code is
called a lot and works as it is. If 'user_data' were to be missing, this
code ought to throw an error (and the error should be a KeyError, not an
AttributeError).