Updated ComboboxEntry to use index for both answer and rawAnswer
This undoes 36ac455 and 0c19009, again removing the validation code
because dropdowns don't allow invalid selections. What those commits
missed is that the combobox needed to be updated to use the index for
rawAnswer. Previously, the combobox was backed by a plain text input,
which meant rawAnswer was text and the entry would look up that text's
index. Now that there's a dropdown, both the validation and the lookups
are no longer necessary. The only logic dealing with text is receiveMessage.