Update cases in parallel using threads
background_queue on prod uses prefork processes (three of them) and
so using threads in the process should work nicely.

Using asyncio would not be faster than using threads, because we are
not using an asyncio equivalent of the **requests** library to send
requests, and so we would have to use threads to wrap requests anyway.