Set local domain backend for gevent workers
Thread-local storage becomes greenlet-local storage when using monkey.patch_all().
https://www.gevent.org/api/gevent.monkey.html#gevent.monkey.patch_thread

Previously system actions were not processed correctly for archived forms, which are copied using _migrate_unprocessed_form(). When the archive_form system action was executed it tried to use the couch backend on a SQL form object, which failed with "AttributeError: can't set attribute" when setting `form.doc_type`.

This was not caught by tests, and indeed is quite challenging to test, because the test runner does not install gevent monkey patches. Running tests with gevent threading patches results in an extremely long lock timeout or a deadlock (not sure which because I didn't wait that long).