Save diffs with original/correct doc type
The hard rebuild on the couch case in diff_cases() may cause its doc type to change to CommCareCase-Deleted, which previously caused new/updated diffs to be saved with the wrong doc type (CommCareCase-Deleted). This could cause unexpected deleted case diffs (which are usually ignored anyway), but it also did not clear away the original CommCareCase diffs, which means diffs for the case remained the same even after it had been patched (confusing) and future patches acted on the original diffs (also confusing, patching did not work as expected).