Remove "really stupid sanity check"
Added in 1de8ae15242ff1a439eaf16acb9e13245659b0dd and 7d0ec99283c53a511ac9fd72782bfbe70809e906, but unfortunately those commits do not have any context on why it was added other than the comment about unit tests. Hopefully this was not a cause of flaky tests, but if it was it should be fixed a different way that only affects test code and not production code.