Remove ineffective admin site login/logout patches
Ineffective because `admin.site.urls`, which references `admin.site.login`, is calculated and installed in the base urls.py before these patches were applied. Tested locally and verified that these were not called. One reason it is safe to remove this is because django-two-factor-auth overrides the admin login pages with a redirect to the main login page.