This feels wrong
This change returns `None` to the API if mappings exist, but they
return no values. The API will presumably respond with a 404. But this
might not be the best outcome. If mappings exist, the administrator
intends to return something. If the mappings result in no values,
that's a problem, but the problem lies with the administrator, not the
user. The correct outcome should be a ConfigurationError.