remove repeated calls to get_case_properties
For each case type a call to `get_case_properties` was made
and for each of those calls the entire set of properties for all case
types is generated resulting in a lot of duplicate work for apps
with lots of case types and lots of modules. Each call to `get_case_properties` could
take a few seconds.