Remove auditcare.views.audited_views
This view would try to load every navigation audit event from Couch, which is an absurdly large amount of data to load in a single request. Needless to say it is impractical and might even have a bad effect on Couch if it were run on a cluster with a lot of auditcare data.