Increase timeout for Repeater.by_domain to 1h
There is good cache invalidation for this function, so we could probably safely go hire.
However I am choosing 1h in case there is a crack in the cache invalidation,
so that the result can never possibly be more than an hour old.
This change is motivated by wanting to further decouple form submissions and case changes
from couch. This should reduce to a tenth the proportion of submissions that need to
make a request to this couch view.