Fix cache invalidation for stale repeaters query
This makes Repeater saves, deletes, etc. eagerly
wait for the view to refresh before then
invalidating the caches for Repater.by_domain with stale_query=True.
This makes stale query actually entirely consistent
but possibly for my inability to deeply understand race conditions.
It also makes tests pass. Because the view is refreshed before
the cache is invalidated, a Repeater save should propagate
near-immediately–however long it takes to refresh the view
and invalidate the cache.