cache enabled toggles for user / domain
Each request to Formplayer calls the `user_details` endpoint which
responds with the list of enabled toggles and previews for the user
and domain. Prior to this PR that involved fetch each toggle out of
Redis at least once and in the worst case twice.

This PR aims to improve the performance of that endpoint.
Locally the difference is 135ms vs 2ms (364 redis requests vs 3)
for the portion that's compiling the list of enabled toggles & previews.

This will also reduce request load on redis (as shown above) at the
expense of adding more data into redis. This only stores the names
of the enalbed toggles for a domain / user instead of ALL toggles names
with their enabled state.