vellum:cfdf7bfea80fac5268e1d9dba8bde60a6832c88e manifest.txt
This is approximate because it was built using that version of vellum, but the resulting main-components.js file, which is not included in this commit, is slightly different. Ideally it would be exactly the same; I'm not sure why it changed. In any case, this is better than what we had before (nothing), and at least a good starting point for tracking what libraries are changing when vellum is updated.