Address review comments.
(The support for reporting all errors in a row required adding and handling a new exception type that is used to collect and raise together the whole set of exceptions gathered during processing of the row. I can't say I'm too happy with this but it seemed the best way to achieve the desired result without a major restructure of the way these errors/exceptions are currently handled in the importer code. Is there an alternative way I couldn't think of?)