Fix test for USE_PARTITIONED_DATABASE=yes
Not sure how this test was ever passing in the past with USE_PARTITIONED_DATABASE=yes and a multi-shard config since AutomaticUpdateRule.iter_cases is called for each shard. Returning a list on a multi-shard config means the same case would be processed for each shard db. Returning an iterator causes it to be exhausted on the first shard, and all other shards to yield no cases.