Beef up error catching/reporting on data dictionary upload/import.
Catch when a case type is missing valid values, and let the user know instead of producing a 500 error
Notice when a case property cell is empty, and let the user know
Notice when a property is listed as having valid values but not found in the base case type information, and let user know.

Also added some basic unit tests of the data dictionary export/import. Import tests for these error situations, export is a bit more basic (so far just tests the right sheets are in the exported file), but it should be easy enough to add on as the data dictionary is worked on more.

Refs QA-3457